+++ packages/server/src/api/controllers/Sales/CreditNotes.ts
@@ -233,7 +233,7 @@ export default class PaymentReceivesController extends BaseController {
       check('entries.*.index').exists().isNumeric().toInt(),
       check('entries.*.item_id').exists().isNumeric().toInt(),
       check('entries.*.rate').exists().isNumeric().toFloat(),
-      check('entries.*.quantity').exists().isNumeric().toInt(),
+      check('entries.*.quantity').exists().isNumeric().toFloat(),
       check('entries.*.discount')
         .optional({ nullable: true })
         .isNumeric()
@@ -755,9 +755,8 @@ export default class PaymentReceivesController extends BaseController {
     const { tenantId } = req;
 
     try {
-      const data = await this.getCreditNoteStateService.getCreditNoteState(
-        tenantId
-      );
+      const data =
+        await this.getCreditNoteStateService.getCreditNoteState(tenantId);
       return res.status(200).send({ data });
     } catch (error) {
       next(error);
