+++ packages/server/src/api/controllers/Sales/SalesEstimates.ts
@@ -172,7 +172,7 @@ export default class SalesEstimatesController extends BaseController {
       check('entries').exists().isArray({ min: 1 }),
       check('entries.*.index').exists().isNumeric().toInt(),
       check('entries.*.item_id').exists().isNumeric().toInt(),
-      check('entries.*.quantity').exists().isNumeric().toInt(),
+      check('entries.*.quantity').exists().isNumeric().toFloat(),
       check('entries.*.rate').exists().isNumeric().toFloat(),
       check('entries.*.description').optional({ nullable: true }).trim(),
       check('entries.*.discount')
@@ -562,9 +562,8 @@ export default class SalesEstimatesController extends BaseController {
     const { tenantId } = req;
 
     try {
-      const data = await this.saleEstimatesApplication.getSaleEstimateState(
-        tenantId
-      );
+      const data =
+        await this.saleEstimatesApplication.getSaleEstimateState(tenantId);
       return res.status(200).send({ data });
     } catch (error) {
       next(error);
